pyproject.toml
[project]
name = "test-opik"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.128.0",
    "google-adk>=1.18.0",
    "openai>=2.15.0",
    "opik>=1.9.85",
    "polars[deltalake,pydantic]>=1.37.1",
    "strands-agents-tools>=0.2.19",
    "strands-agents[anthropic,gemini,mistral,otel]>=1.22.0",
]


[tool.uv]
dev-dependencies = [
    "ruff>=0.14.13",
    "fastapi-cli>=0.0.20",
]
